Cập nhật lớn: Hoàn thiện pipeline ML, Discord bot chống trùng lặp, dữ liệu Binance thống nhất
- Thêm các thuật toán ML đầy đủ: Linear/Logistic Regression, KNN, Random Forest, Clustering
- Triển khai pipeline huấn luyện production với export model và quick_loader
- Xây dựng Discord bot với nguồn dữ liệu realtime Binance thống nhất và cơ chế chống trùng lặp
- Thêm web dashboard tích hợp Flask/Plotly
- Cải thiện cấu hình Docker Compose với các dịch vụ web/bot/demo
- Nâng cao thu thập dữ liệu với tích hợp MongoDB
- Thêm các script test và validation toàn diện
- Triển khai xử lý lỗi và logging mạnh mẽ toàn hệ thống

+++ app/ml/metrics.py
@@ -206,13 +206,13 @@ class CryptoMetrics:
     def _calculate_mape(self, y_true: np.ndarray, y_pred: np.ndarray) -> float:
         """Mean Absolute Percentage Error"""
         mask = y_true != 0
-        return np.mean(np.abs((y_true[mask] - y_pred[mask]) / y_true[mask])) * 100
+        return float(np.mean(np.abs((y_true[mask] - y_pred[mask]) / y_true[mask])) * 100)
     
     def _calculate_smape(self, y_true: np.ndarray, y_pred: np.ndarray) -> float:
         """Symmetric Mean Absolute Percentage Error"""
         denominator = (np.abs(y_true) + np.abs(y_pred)) / 2
         mask = denominator != 0
-        return np.mean(np.abs(y_true[mask] - y_pred[mask]) / denominator[mask]) * 100
+        return float(np.mean(np.abs(y_true[mask] - y_pred[mask]) / denominator[mask]) * 100)
     
     def _directional_accuracy(self, y_true: np.ndarray, y_pred: np.ndarray) -> float:
         """Accuracy of predicting direction of change"""
